refactor: isolate CSV header validation
